fix: use utils.NewHTTPClient in fetchBillText, move TestTruncate to templates package
Agent-Logs-Url: https://github.com/philspins/open-democracy/sessions/4f4f929a-711c-45fa-8d49-f2fd94325a9f
